feat(preprocessing): Update imputer to support generic numeric types
This commit enhances the SimpleImputer class to work with generic numeric
types by:

1. Changing the internal statistics storage from f32 to f64 for better precision
2. Making the fit/transform methods generic to accept any numeric type that can
   be converted to f64
3. Updating type signatures in the Transformer trait implementation
4. Replacing Box<dyn Transformer> with TransformerType enum in ColumnTransformer
   for better type safety

These changes allow the imputer to handle various numeric types (i32, f32, f64, etc.)
while maintaining precision through f64 internal representation. The transform
output now consistently returns Array2<f64> regardless of input type.

+++ src/parameters/preprocessing.rs
@@ -1,10 +1,11 @@
+
 /// Imputation strategies supported by SimpleImputer
 #[derive(Debug, Clone, Copy, PartialEq)]
 pub enum ImputationStrategy {
     Mean,
     Median,
     MostFrequent,
-    Constant(f32),
+    Constant(f64),
 }
 
 /// Strategy for handling unknown categories in OneHotEncoder
